requirements.txt
flask
requests
coloredlogs
sqlalchemy
pytest
pytest-xprocess
